requirements.txt
python_telegram_bot==11.1.0
requests
